Partially revert 12a83c99 for boost's header warnings
The cited commit caused the include with -isystem of default system
paths when boost is installed in them. This messes with GCC's default
order for system headers, which in turn can cause other issues. In
particular, we are triggering a known GCC issue:

https://gcc.gnu.org/bugzilla/show_bug.cgi?id=70129

The proposed alternative is "don't mess with isystem with system paths",
because apparently it does more than just ignore the warnings originated
in those headers. So let's revert back to including Boost with -I when
it is located in the usual suspects.