Explicitely initialize taskinfo fields
Prevent some unexpected side-effects when uninitialized values in newly
instantiated taskinfos are filled with non-zero values.
Note that for the case of SpawnTask.cpp, this change is a no-op:
value-initialization already took care of this. However, we include it
for consistency and to prevent future issues if we change the code to
not value-initialize the taskinfo.