Fix metadata calculation
Since we allocate an extra pointer on top of the task metadata to keep
track of its location, we have to compute this extra overhead in order
to verify that we don't exceed nOS-V's max metadata size.

Essentially, tasks having more than 4088 and up to 4096 bytes of
metadata were making NODES crash.